You can now create multiple entries for each exchange.
This helps support having multiple accounts on an exchange, or things like sub-accounts on Deribit. You can name each set of API keys as you like now, as long as you also include the name of the exchange. Your commands refer to your given name to identify which set of credentials to use. For example `myBot(BTCUSD) {wait(5) }`, will find a set of credentials in your config with the name ‘myBot’