Rounding and precision is now a lot more sophisticated.
Bitfinex is rounding using significant figures instead of precision, deribit correctly handles the tick size and all the order types use the exchanges rounding settings when calculating prices and amounts for orders.

Also, all the exchanges now look up the symbol when the exchange starts up, so the correct precision rules can be set up for the instrument you are going to be trading on.