Ping pong orders are now better at rounding
as well as handling orders with an order count of zero. The figures passed on to the scaled order are more likely to produce good results.
The lower limit of 2 orders in the ping pong has also been dropped, so you can now create one with 1 order, or zero (though that will do nothing, obviously).