Key data is now stored per symbol
If an exchange is handling multiple trades on different symbols at the same time, some data would not have been in a suitable range. This was common if you wanted to run multiple ping pong orders, or other long-lived orders, on different symbols on the same exchange at the same time.

The exchange is now informed when trading a new symbol is started, so it can start to listen for relevant events.

We also track some data, such as min order size and precision settings separately for each active symbol in the exchange.